+++ R/expect-constant.R
@@ -12,17 +12,17 @@
 #' @examples
 #' expect_true(2 == 2)
 #' # Failed expectations will throw an error
-#' \dontrun{
-#' expect_true(2 != 2)
-#' }
-#' expect_true(!(2 != 2))
-#' # or better:
-#' expect_false(2 != 2)
+#' show_failure(expect_true(2 != 2))
 #'
-#' a <- 1:3
-#' expect_true(length(a) == 3)
-#' # but better to use more specific expectation, if available
-#' expect_equal(length(a), 3)
+#' # where possible, use more specific expectations, to get more informative
+#' # error messages
+#' a <- 1:4
+#' show_failure(expect_true(length(a) == 3))
+#' show_failure(expect_equal(length(a), 3))
+#' 
+#' x <- c(TRUE, TRUE, FALSE, TRUE)
+#' show_failure(expect_true(all(x)))
+#' show_failure(expect_all_true(x))
 #' @name logical-expectations
 NULL